+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - 初始化项目 | Init project
 - 添加项目依赖 `utools-api-types` 与 `utools-helper` | Add dependencies `utools-api-types` and `utools-helper`.
 - 创建功能 `config` | Create a feature `config`.
+- 配置 vite 生产环境打包选项 | Config vite production build options.